test(ui): split device-credential store tests and drop the identity peek cast
Two CI gates on the rebased head:

- check-lint (max-lines): gateway-store.test.ts reached 1053 effective lines
  (cap 1000) — main grew under this branch and the credential tests tipped
  it over. Move the self-contained "stored device credential" describe into
  a colocated gateway-store.device-credential.test.ts with a slim preamble
  instead of suppressing; the original returns to 917 lines (main: 918).

- assertion-safety ratchet: the re-added peekStoredDeviceIdentityId parsed
  with `as StoredIdentity`. Parse to unknown and validate with the file's
  existing isRecord guard so the shape is checked at the storage boundary
  rather than asserted — no new cast, no SAFETY carve-out.
